+++ cmd/ethermint/rebuild.sh
@@ -1,3 +1,0 @@
-rm -r /usr/bin/ethermint
-go build
-mv ethermint /usr/bin