+++ cmd/gelchain/rebuild.sh
@@ -1,0 +1,3 @@
+rm -r /usr/bin/gelchain
+go build
+mv gelchain /usr/bin